requirements.txt
openai
streamlit
langchain
tiktoken
google-search-results
langchain-openai
python-dotenv
requests
watchdog==4.0.0
parea-ai==0.2.95
pillow==10.2.0